divrep oul: allow for forced primes in mintau calculations
Forced primes are always used, so it is not useful to include them in
the list of primes we check for. However they can be used without
filling an odd-factor "allocation", so we need to extend the list
we _do_ check for.

Allowing for that means we should avoid the fatal "no mint match found",
and at runtime we should be slightly faster (since we won't waste time
checking for bit-vector matches that can never match).