divrep: cast out squares in gen_pell
gen_pell() could be unreasonably slow with large D, since the continued
fraction could have a very long period. However if called on (D, N) with
D = a^2 b, we can simply replace it with (b, N) and use a linear filter
to divide the resulting y-values by a.

This improves a case such as (15 . 23^9, 10) from 45 minutes to first
solution to instant.