divrep oul: default forced prime to k
Treat './oul n k' as equivalent to './oul -fp<k> n k', since this seems
to be the useful setting in pretty much all cases.

We retain -fp as a way to override this for now, but if we can make this
mandatory that will allow a lot of simplification - recurse() will no longer
need to check for the need to try_force, v->{forced} will not be needed,
the -f option also becomes redundant.