divrep oul: refine forced primes
For n == 2 (mod 4), allocating even p^1 is useful as it forces a square.
For n == 0 (mod 4), it is not, and we end up wasting a bunch of effort
allocating individual p^^x at different locations which could be better
handled implicitly via recurse().

Accordingly, when n == 0 (mod 4), we elide from the list in $pforce{$p}
any batch that allocates to only one element, replacing those with an
empty placeholder; we additionally set $maxforce[$i] and use it to
decide which prime to start looping from in recurse() so that we handle
the newly elided cases there without duplicating those that remain.

For f(4n,k) with k prime this introduces the slight inefficiency of
a forcing set consisting only of the empty placeholder; this should
be avoided if we can find a cleaner way to handle it.

It may also be worth forcing more when n == 4 (mod 8), since even p^1
is somewhat constraining in that case.