A090252: first pass at third attempt
Previous attempts tended to run out of memory; however some new awareness
can mitigate this.

First, the vast majority of values are the primes in increasing order.
So it is enough to maintain a hash of (n => a(n)) where a(n) is composite,
and a pair of values to record m(n), the next prime we will use for an a(n),
and the next prime that will become newly available to use as a factor.

Second, the vast majority of available primes can be expressed as a single
range of consecutive primes sqrt(m(n)) < p < m(n); up to n=2.5 x 10^7,
it would seem we never have more than 3 primes available outside that
range.

This initial attempt attempted to separate the small primes < sqrt(m(n)),
expecting them never to number more than 2, from a heap of larger
primes > sqrt(m(n)) not part of the contiguous range, but found a) that
this fails at n=99331, when we first see 3 small primes, and b) that
all the effort in maintaining the heap was pointless since it did not
exceed 1 element up to n=2.5 x 10^7.