A090252: fix the third attempt
This now takes zero to two arguments to request an output style:
- 'a090252' to produce an a-file for A090252
- 'a354164' to produce an a-file with just the composites
- 'S <k>' to show only those a(n) divisible by <k>
- default shows additional diagnostic information of available primes.

We no longer use a heap, we just have a sorted list of "small" primes
that are available, and the endpoints of the range of consecutive
primes > sqrt(m(n)) for the rest.

The code as a whole is much simpler, and uses near-constant memory -
the only (barely) noticeable growth comes from the lookups for composite
values, of which there are 10^4 in the first 10^8 values.

This took about 4 minutes on my machine (with load average 7) to find
a(10^8).

The code uses native integers where possible, spilling to Math::GMP
objects otherwise. That seems unlikely to be useful though, so the
code could be made yet simpler and faster by removing the Math::GMP
support.