divrep inject: tool to check and commmit a sequence
Usage: ./inject [-y<type>] [-l<hlim>] [-n] [-f] n k d

Default behaviour is to check for a sequence oneseq(n, k) = d, and record
the result in the database as an upper limit if found to be correct.
It uses partial factorization to speed up the checks.

-y<type> can be used to override the sequence type; -l<hlim> sets
the limit for trial division when factorizing; -n is used to skip
the database (effectively a dry run); -f is used to mark the run as
"full" rather than "partial": if a result is found it will be marked
as minimal rather than as an upper limit.