divrep/oul: new option -f<force_all>
For n == 2 (mod 4), we force every set of possible assignments for
each p <= k, since even a single power provides a powerful constraint,
by forcing the affected value to be a square. For n == 0 (mod 4), we
force only those cases where two distinct values are divisible by the
same prime, and leave recurse() to assign the rest.

However for n == 4 (mod 8) a single prime can still provide a useful
constraint. Forcing all assignments for each p < k, for example, is
a net win for oul(12, k) when k is 5 or 7 (but not when k is 3, 4 or
6): counting the number of recurse and walk iterations without and
with such force shows:

oul(12, 3): 29/42 v. 34/47
oul(12, 4): 274/1008 v. 309/1218
oul(12, 5): 918/4668 v. 762/3975
oul(12, 6): 27796/1035672 v. 31694/1106332
oul(12, 7): 423664/72917334 v. 392804/61253929

The last case ran 12.7% faster overall with force.

So it seems plausible that setting -f to k-2 will be a win; certainly
it is worth experimenting with.

This must currently be set the same on each restart; it should be
included in the 001 log lines to avoid the need for that.