divrep/oul audit: allrootmod, allsqrtmod return bigints
There has not yet been a release of Math::Prime::Util with allrootmod
and allsqrtmod, though they have been available in github for a while.
However MPU functions don't guarantee to return bigints.

Ensure that the versions we expose are guaranteed to return bigints,
even when we can take advantage of a new MPU release.