divrep coul: check overshoot on forced primes
For unforced primes we set a limit in advance, but for forced primes
we do not; so check after each allocation whether the resulting q_i > max.
It isn't clear that it saves much even when it triggers, but setting
forced primes is a rare event in any significant run so the cost of the
checks is small even in the worst case.