divrep coul: checkme for semiprimes
The TODO item 'test "other" tau cases in parallel' involves unrolling
the individual test strategies from factor_one(). Once we have that,
it is worth considering whether it is useful to handle the t=4 case
separately, and to see how we should weave that in.

Note that factor_one comments that QS (_GMP_simpqs()) is "fantastic if
it is a semiprime, but can be slow and a memory hog if not (compared
to ECM)". However it is not used in the is_semiprime() implementation:
that tries pbrent, pminus1 and several ECM runs before giving up and
falling back to default factorization (which will redo most of those).