divrep coul: extend -g option to ratio
We calculate the maximum prime limp that we may need to allocate at
a particular level, and use cost = (limp - startp) as a proxy for the
cost of recursing at this level.

We also calculate 'walk', the number of iterations we expect to
perform if we stop recursing at this level and just walk_v() all
the values at this point.

By default, we short-circuit the recursion with an immediate call
to walk_v() when walk < cost. Setting gain via -g allowed us to
defer the short-circuit by changing the calculation to be
walk * gain < cost.

We now extend -g to take two values, which act as a ratio:
"-g<antigain>:<gain>", and change the calculation again to
(walk * gain) / antigain < cost, for more fine-grained control.

<gain> and <antigain> are both treated as 1 if supplied as 0, or
as the empty string.

Gain can be set with "-g20" as before, or with "-g:20". Anti-gain
can be set with "-g20:"