divrep find-constraints
Given a set of parameters for gtauseq (-nx, -cp, -c, <n>, <k>), analyses
the debug output to produce a deduplicated list of the different types
of constraints found.

This is primarily intended for the case where (n, k) is invalid, to show
the logic by which that has been determined.

TODO: extend this a) to minimize the set; b) to discover max(k) by running
bisect-g to find a bound, then bisecting the option parameters to reduce
the sets before proceeding to full minimization.