divrep coul: option -s to control random seed
MPUG code forces a random seed of time(); override this first by
immediately resetting it to have seed=1, then agin to the chosen
seed if selected with -s.

This makes runtimes easier to reproduce; however, times may not be
consistent between runs if factorizations do not happen in the same
order. I don't have a good solution for that right now - it is not
practical to reset the seed before every factorization.