divrep coul: optionally test "other" cases in parallel
If built with -DPARALLEL, after checking values required to be prime
or square, we collect the remaining values to be checked so that they
can be checked in parallel: we do the fastest checks on each first,
since the first failure allows us to abort the whole procedure.

This breaks the tests used by factor_one() into a list of 41 individual
functions; a future enhancement could let us specify how far through the
list to go before displaying the candidate as a "maybe" (via the '-o'
option).

Note that this currently skips the power check previously done immediately
after trial division; we probably need to restore that (at the cost of
a bit more complexity).

The Makefile now builds both './coul' as before, and './pcoul' with
-DPARALLEL.