divrep coul: sort "other" cases in walk_v and tau_multi_run
We sort first on highest power of 2 dividing t (ascending), then on
t (ascending) then on n itself or the qq multiplier (ascending).

Additionally sorting the list of inverses or the list of prime cases
seemed to make things slower: more investigation needed. (Sorting the
inverses seems like it should be a no-brainer.)