divrep coul: bugfix in _allrootmod_prime_power_r()
In translation from perl code a section was missed for the case k==p:
when lifting from p^e_1 to p^e_2, we must lift r to r(1 + jp^{e_2-1})
for j in 0 .. p-1.

This causes various roots to be missed, eg we will return (1) for the
square roots of 1 (mod 4), instead of (1, 3).