divrep coul: experimentally account for 2^{x+1}
When we allocate 2^e at v_i, set v_0 = 2^e - i (mod 2^{e+1)) immediately
in CRT rather than relying on an inverse check in walk_v() to disallow
cases that would leave a higher power of 2 dividing v_i.

TODO: check if there are any other knock-on effects we have not
accounted for.