divrep coul: new options '-a', '-b<i>' for sharding
With -a, show the valid sets of forced-prime allocations with indexes,
then exit.

With -b<i>, process normally but only for the set of forced-prime
allocations with index <i>.

Note that short-circuiting means -a will still fully test (and not
show) any allocation that leaves t=1. This is assumed to be a good thing.

The output of -a is a new "203" log line, showing the shard index, the
set of allocations, and the number of forced squares (if non-zero).
This shows, for example, that for D(12,15) with -f13 we have 488 shards
without forced squares, 75 shards with a single forced square, and 688
with multiple forced squares.