divrep coul: fix and improve batch overshoot test
We would reject a batch if v_i.q > max, but we should have been checking
if (v_i.q - i) > max. We were also uselessly checking the same thing
again after applying each secondary, which would have exacerbated the bug
if we had been checking against the correct value.

If we're going to check that, we can do an even better job by checking
level.rq > max (as we would with CHECK_OVERFLOW), so use that instead.