divrep coul: fix apply_batch bugs
1) before we call walk_1(), we must have applied a complete batch so
that we can correctly suppress bad powers;
2) update_residues() needs a lot more effort to handle the case in
which a secondary element of the batch hits the known square - we
must first tell it about the secondary, so that the residues are
divided down (and potentially upgraded for higher "g"), then tell
it about the primary (represented as the _additional_ power over
the secondary) so that the residues are appropriately extended.

To achieve this, we remove the handling of both cases from
apply_primary/secondary and instead treat them separately at the
end of apply_batch. update_residues() no longer takes an is_secondary
parameter - we simply avoid calling it for the case that was meant
to protect against - but now takes a 'retry' parameter representing
the power supplied for the secondary when calling a second time
for the primary.

Net cost is that update_residues() is a touch slower for the extra
checks; all other changes should only affect the rare case of batch
allocations, or even improve things slightly.