divrep coul: new #define LARGE_MIN
Improve handling of large minimum value (with -x<min>:) - take account
of min when calculating the recurse/walk tradeoff; check in walk_v()
that the starting value of the iterator fits in 64 bits.

For the majority case where min is 0 or small, we assume these checks
are an unnecessary expense; hence we do not enable them by default.