divrep coul: temporary bugfix for batch recovery
When invoked with -b to run a single batch or a range of batches,
we rely on counting the batches as we see them to determine the
current batch number. However on recovery we inject the current
state directly, so we won't know the current batch number.

As a temporary fix, which works at least for a single batch, detect
this case and assume the current batch is the requested batch.

The full fix needs a wider rethink, probably to replace the batch
counter with a direct representation of the pattern this batch
represents, as an array of t_forcep.batch indices.