eulerian: allocate all forced edges
We keep a "pending" stack of any forced edges, and allocate all of them
before looking for optional ones to allocate.

To know which edges are forced when a vertex has had all the allocations
it is allowed in one direction, we attach a list of its edges to each
vertex.

We replace the 'more' on a stacked allocation that signalled whether
it was the first of a two-way choice with a separate stack of choice
points. So now derecursion involves unwinding the stack to the last
choice point.