topsu initial commit
Starting point to investigate the class of topologically distinct sets
of possibilities for a number in an incomplete n^2 x n^2 sudoku puzzle,
with weight w more possibilities than required.

The starting point is code to find a canonical form for the w^3
incidence matrix; this proof of concept traverese the full 2 (w!)^3
possibile rearrangements, which takes about 60s for a w=5 example.