topsu: faster canonicalization
Scan first for rows or columns in which some region has maximal 1's:
we know such a maximal set must form the beginning of the first row
in the canonical representation. Restrict choices of x0/y0 to those
which pick a point from one of these maximal sets.

This reduces time to canonicalize a sample size-9 matrix from 1783s
to 322s.

A further refinement is possible by splitting regions with maximal
lines into those in which at least one other region has further 1's
on the line, and those in which no other region does.