topsu: canonicalize regions
We can reduce the number of sets of regions we consider with an
additional phase of canonicalization.

This reduces the number of full positions to consider for a(6) by 20%,
but saves little if any time since a set of full regions is worst-case
for the Topsu canonicalization process; however a dedicated full-region
algorithm can be much faster.