topsu: detect simplest, most expensive case
We'd like to detect cases where some of the bands or stacks are
topologically identical, but that seems hard. However the worst case -
a single band of size w, such that the points (perforce) form a simple
cycle like:
  aa..
  b.b.
  .c.c
  ..dd
already gives us a big chunk of the savings: 16x faster at n=6, which
suggests n=7 could complete within 16 hours.