topsu: expand NestedLoops call in assign_iter
In preparation for further intervention, replace the NestedLoops()
call that creates the iterator to find each arrangement of the spares
across the regions with the necessary actual code.

This already gives some small savings (at w=6 about 16% of assign_iter,
about 4.5% overall), but will allow insertion of short-circuiting checks
that should make more of a difference.