hypersq: let v_0-v_1 set minimum distance
Require all points selected to differ from all other points in at least
as many places as v_0-v_1.

If any pair v_i-v_j violates this, the resulting set can be transformed
into a lexically earlier one that we would already have tested by
xor with v_i, then reordering dimensions to move the bits set in the
image of v_j to be in the least significant positions.