hypersq: detect simple 2D symmetries
The set { 0, 1, 2^k+1 } can be tranformed by xor 2^k+1 and reordering
of dimensions to { 0, 1, 2^k }, which would already have been checked.

This is a simple case of a much larger class of such transformations,
but detecting all of them would require full canonicalization, and it
is unclear how fast we can make that.