hypersq: bug fix, must clear low bits when setting high bits
A combination has invalidated earlier assumptions, so ensure a) that
when forcing extra bits due to dimensional grouping we clear any lower
bits; b) that when changing the point to consider we loop back to
allow earlier checks to be performed again. These changes mean we now
also need an extra check (sp > 1) when looking for 2D reflections.