hypersq: add full canonicalization
Using this for every new point is horribly slow, but demonstrates that
the number of unique sets matches A039754.

New option '-tc$a,$b,$c...' specifies that we should canonicalize
only for the $a'th, $b'th, $c'th ... points of the set. It turns out that
if we use it only for points up to the floor(n/2)th, we actually go faster,
so that is the default selection.