unitfraction: test for 4n+3 divisors in numerator
The sum of two square unit fractions cannot have a numerator divisible
by any prime factor of the form 4n+3; we can save a lot of effort by
checking for this. However, unlike the case for the denominator, we
have no constraint on how hard the numerator may be to factorize.
So we aim for a balance by conducting only a quick test on the lower
bits.

In tests, this reduced the work done in find_square_s2() by a factor of 9.
Full factorization would have given a further 1.6x improvement, but at
unknown cost.