unitfraction: bugfix for squarefree denominator
When factorizing the denominator for `minset -q`, we shift any odd
powers into a spare multiplier which should always be applied; however
when that left no other factors, we failed to apply the extra multiplier.
This gave wrong results for example with `minset -q 2 17`.

(This would rarely affect results except for square-free denominators
with k=2.)