fermant: expected value of cheapest path across random grid
Consider a rectangular grid of a x b nodes, in which the cost of traversing
from one node to an orthogonally adjacent node varies uniformly over [0, 1].
What is the expected cost of the cheapest path from the southwest to the
northeast corner, where each step moves only eastwards or northwards?

The perl program mkint calculates f(a,b) by finding ways to split the
integrals such that the min() is known and thus integrable.

The code is currently too slow to calculate f(3,3) - where f(5,2) splits
the base integral into 11,852 parts, for f(3,3) we will have (estimated)
between 1e7 and 1e10 parts. The plan is therefore to rewrite it in C for
speed, with progress logged to file so that we can recover after any
interruption.

The perl program mkstrat uses the same approach to generate lines of
a proof for the calculation, suitable for checking by github.com/hvds/axiom
(but needs access to modules within that codebase).