fermant: initial C version
This seems to work for g(2,n); for g(3,3) we crash at around 11GB with
some 81M fragments, possibly because the stack and heap are meeting each
other: the segfault complains that limit_lc(96665989) is unreadable, even
though the last limits realloc to size 144998983 appears to have succeeded.

Next step is to turn the splitting phase into a process of reading from
one data file and writing to another, and the same for integration.