fermant: bugfix, don't skip terms for distrib
If a distrib generated a term exactly cancelling out an existing one,
removing the resulting zero would swap in the last term of the expression
to fill the gap. When the removed term was at an offset already passed
over in the process and the swapped-in term needed distributing, we would
fail to see it, causing incorrect results.

We fix this by specifically requesting a denormalised result when adding
terms in this context, and applying a new renormalisation pass when we've
finished the sweep.