vdwaerden/analf: analyse pruning power
Analysing the results seems to show that the number of iterations required
is reasonably well predicted as a power of n (though obviously this is not
stable). Capturing that info will help us better understand the trade-off
when doing more expensive work for better pruning.

The analysis shows iter^(1/n), iter^(1/f(n)) along with n, f(n) and the
number of consecutive failures since the last improvement.