cf: pruning by chunks
Rather than just count allowed bits in the blocked vector, slice it into
chunks and sum the largest possible subsets from the allowed bits in each
chunk.

We use a chunk size of 16 bits, and prepare two 64k lookup tables for the
purpose: one that just records the cardinality of the largest possible
subset, and another that includes a mandatory (virtual) 17th bit that
must always be included - the latter takes advantage of our requirement
that any maximal subset must always include both 1 and n.

This runs about 2x faster to n=143, and saves variously 6 to 11 parts
per thousand in the iteration exponent. (For some reason the exponent
improvement seems to drop off as n increases - perhaps the extra
effectiveness of the "mandatory 17th bit" is diluted by additional
chunks?)