cf: generalize chunk size
24-bit chunks reduce exponent by another 3 parts per thousand, and
improve overall speed by 1.1x for a 2-3s startup time, and 32MB of data.

Timings are modified to treat the point after initialization as t_0.
Removed unused reverse_16().

Note that for n in (29, 30, 34) the number of iterations is slightly
worse; that does not appear to be a bug, just a (rare) artifact of how
the chunks split.