vdwaerden interim checkin: work chunkwise
Provide an iterator from allsub to walk through all valid subsets;
in cf, initialize lexically sorted lists for each width up to the
chunk size, and in the recursion proceed a chunk at a time.

This is horrendously worse than what we had before. In part that is
because when, for example, the first bit of a new chunk is masked,
the previous implementation could immediately short-circuit, while
the new implementation must still walk through all the chunks.

It remains to be seen if we can refine this to the point it actually
improves on the previous implementation.

Note that while we ostensibly support chunks of 8, 16 or 24 bits,
cross_product() on 24 bits fails since it tries to return a 71-bit
result in a 64-bit variable.