cf: prep all 8-bit lookups
We prepare lists of every set of valid subsets we might need: accounting
for excluded bits (where two preceding bits conflict with a single bit
in the value), reverse-excluded bits (where a preceding bit conflicts
with a pair of bits in the value), for required minimum number of
elements, for restricted length (in the final chunk), and for forced
inclusion of 1 and n in the subset.

This requires up to 3e6 lists each up to 106 bytes in length.
Extending this would require exponentially more memory: 18 GB for
10-bit lookups, 1 TB for 12-bit and 4.8 PB for 16-bit.

We do not yet take advantage of any knowledge of required minimum
number of elements, that should be the final piece of this puzzle.