cf: use bitcount
With (c + xc + y) bits to go, where c is the chunksize, we check each
possible value of y against the recorded largest subset possible, and
combine with known largest subset with given mask for the x chunks
following the current chunk, to determine the minimum number of bits
we must include in the current chunk.

The result is that we have fewer iterations for a handful of n, more
iterations for the majority of cases, and since we are doing more work
on each iteration the result is about 4x slower.

Conclusion: this is a failed experiment, but it is possible we can
use what we learned to make some further improvements to the original
code.